requirements.txt
django
psycopg2-binary
python-dotenv
djangorestframework
drf-yasg
streamlit
openai
pdfplumber
python-magic
python-docx